refactor(tags): use new reusable hero
